# Optimize regexp handling for width